Move AttrDict under beets.util
